move some repeated functions to a new util.py file
