+++ diffmap/codiffmap.py
@@ -29,97 +29,14 @@
 # -- analysis
 # 16. peak_amplitudes
 
-
+# is it ok to use "from .utils import *" here or should I preserve the
+# namespace for the utility functions? Going to do the latter...
 import numpy as np
 import copy as copy
+import .util
 
 
 # ---- data processing
-
-
-def states_to_mri(data, p0=0, offbool=(False, False), invert_sin=False):
-    """
-    converts states-like data to mri-like form. assumes that you have already
-    done any zero-filling and apodization that you want to do, but allows you
-    to apply a global zeroth order phase shift if you want.
-    """
-    # First, get the array dimensions we'll need for the calculation.
-    # Because the initial array is hypercomplex in t1, the final number of
-    # t1 points, N1_f, is the same as the initial number, N1_i. We have to
-    # double the number of t2 points, though, so N2_f = 2*N2_i.
-    N1_i, N2_i = data.shape
-    N1_f = N1_i
-    N2_f = 2 * N2_i
-    N1mid = N1_f // 2
-    N2mid = N2_f // 2
-
-    offbool1, offbool2 = offbool
-
-    # Apply a zeroth order phase shift, if desired
-    data = ng.proc_base.ps(data, p0=p0)
-
-    dataA = np.zeros((N1mid, N2_i))
-    dataB = np.zeros((N1mid, N2_i))
-    dataC = np.zeros((N1mid, N2_i))
-    dataD = np.zeros((N1mid, N2_i))
-    # If we think of the hypercomplex input data as two different arrays,
-    # S_cos(t1, t2) and S_sin(t1, t2), interleaved along t1, then we can
-    # break it down into components as follows:
-    #     dataA = S_cos(|t1|, |t2|).real
-    #     dataB = S_cos(|t1|, |t2|).imag
-    #     dataC = S_sin(|t1|, |t2|).real
-    #     dataD = S_sin(|t1|, |t2|).imag
-    # These arrays are all of shape (N1mid, N2_i).
-    dataA = (data[::2, :]).real
-    dataB = (data[::2, :]).imag
-    dataC = (-1)**(invert_sin) * (data[1::2, :]).real
-    dataD = (-1)**(invert_sin) * (data[1::2, :]).imag
-
-    data_out = np.zeros((N1_f, N2_f), dtype=np.complex)
-
-    # Construct data_out one quadrant at a time. How this works depends
-    # on offbool in a slightly complicated way, so build the slice
-    # objects for each quadrant first. There's probably a fancier way
-    # to do the logic here (esp. for q3), but if so, I don't know it.
-
-    # QI (t1>=0, t2>=0) is always the same:
-    q1 = np.s_[N1mid:, N2mid:]
-
-    # QII (t1>=0, t2<=0) depends on offbool2:
-    if offbool2:
-        q2 = np.s_[N1mid:, N2mid - 1::-1]
-    else:
-        q2 = np.s_[N1mid:, N2mid:0:-1]
-
-    # QIII (t1<=0, t2<=0) depends on offbool1 AND offbool2:
-    if offbool1 and offbool2:
-        q3 = np.s_[N1mid - 1::-1, N2mid - 1::-1]
-    elif offbool1:
-        q3 = np.s_[N1mid - 1::-1, N2mid:0:-1]
-    elif offbool2:
-        q3 = np.s_[N1mid:0:-1, N2mid - 1::-1]
-    else:
-        q3 = np.s_[N1mid:0:-1, N2mid:0:-1]
-
-    # QIV (t1<=0, t2>=0) depends on offbool1:
-    if offbool1:
-        q4 = np.s_[N1mid - 1::-1, N2mid:]
-    else:
-        q4 = np.s_[N1mid:0:-1, N2mid:]
-
-    data_out[q1] += dataA - dataD + 1j * (dataB + dataC)
-    data_out[q2] += dataA + dataD + 1j * (-dataB + dataC)
-    data_out[q3] += dataA - dataD + 1j * (-dataB - dataC)
-    data_out[q4] += dataA + dataD + 1j * (dataB - dataC)
-
-    # in the offboolN = False case, we end up double-counting the
-    # tN=0 data, so we have to divide that row/column by 2 to fix it
-    if not offbool1:
-        data_out[N1mid, :] /= 2
-    if not offbool2:
-        data_out[:, N2mid] /= 2
-
-    return data_out
 
 
 def fft_phase_1d(origdata_2d):
@@ -153,27 +70,6 @@ def fft_phase_1d(origdata_2d):
     origdata_semifft_ph = origdata_semifft * pcorrmat_f2
 
     return origdata_semifft_ph
-
-
-def get_phasecorr(N, offbool):
-    """
-    Creates a phase correction array, to make the FFT of a Hermitian
-    array all real. **Only works in 1D right now!**
-    'N' is the length of the vector
-    'offbool' is a boolean that controls whether there's a dt/2 shift
-    in the time values:
-    - if time[N/2] is t = 0, set offbool = false
-    - if time[N/2] is t = dt/2, set offbool = true
-
-    Returns a 1D vector 'phasecorr' of length N.
-
-    This function assumes an even number of points, so it's yet not as general
-    as it could be.
-    """
-    # phasecorr = np.exp(-1j*2*np.pi*(N/2-0.5*offbool)/N*(N/2-np.arange(N)))
-
-    return np.exp(-1j * 2 * np.pi * (N / 2 - 0.5 * offbool) /
-                  N * (N / 2 - np.arange(N)))
 
 
 # ---- data preparation
@@ -353,7 +249,7 @@ def p1_proj(data_t, mask, offbool):
     data_f = np.fft.fftshift(np.fft.fft(data_t))
 
     # get the phase correction wave
-    phasecorr = get_phasecorr(len(data_f), offbool)
+    phasecorr = util.get_phasecorr(len(data_f), offbool)
 
     # phase correct input data and throw out the imaginary part
     data_f_ph = (data_f * phasecorr).real
@@ -611,7 +507,7 @@ def peak_amplitudes(data, peak_list, grid_bool):
     # correction
     N1 = data.shape[-1]
     amp_list = peak_list.copy()
-    pcorr_f1 = get_phasecorr(N1, 1)
+    pcorr_f1 = util.get_phasecorr(N1, 1)
 
     # we output a peak for each 2d spectrum: peak_amp[spectrum][peak]
     if Nspectra == 1:
